pages 9,10 responsiveness
